feat: move key routes to root other root stack
