+++ apps/mobile/src/routes/main/browse/BrowseStack.tsx
@@ -7,9 +7,7 @@ import {
 } from './browse.interfaces';
 import { useTranslate } from '../../../i18n/useTranslate';
 import { defaultHeaderOptions } from '../defaultHeaderOptions';
-import { defaultSubHeaderOptions } from '../defaultSubHeaderOptions';
-import { SeriesStack } from '../series';
-import { BrowseScreen, NativeVideoPlayerScreen } from '../../../screens';
+import { BrowseScreen } from '../../../screens';
 
 const StackAuthorized = createNativeStackNavigator<BrowseStackParamList>();
 
@@ -27,25 +25,6 @@ export const BrowseStack = () => {
           }),
           animation: 'slide_from_right',
         })}
-      />
-      <StackAuthorized.Screen
-        name={BrowseStackScreenNames.SeriesStack}
-        component={SeriesStack}
-        options={() => ({
-          ...defaultSubHeaderOptions({}),
-          headerShown: false,
-          presentation: 'modal',
-        })}
-      />
-      <StackAuthorized.Screen
-        name={BrowseStackScreenNames.NativePlayer}
-        component={NativeVideoPlayerScreen}
-        options={{
-          ...defaultSubHeaderOptions({}),
-          headerShown: false,
-          autoHideHomeIndicator: true,
-          presentation: 'fullScreenModal',
-        }}
       />
     </StackAuthorized.Navigator>
   );
