+++ apps/mobile/src/routes/main/browse/browse.interfaces.ts
@@ -4,31 +4,13 @@ import { SeriesStackParamList } from '../series';
 
 export enum BrowseStackScreenNames {
   Browse = 'Browse',
-  SeriesStack = 'SeriesStack',
-  NativePlayer = 'NativePlayer',
 }
 
 export type BrowseStackParamList = {
   [BrowseStackScreenNames.Browse]: undefined;
-  [BrowseStackScreenNames.SeriesStack]: NavigatorScreenParams<SeriesStackParamList>;
-  [BrowseStackScreenNames.NativePlayer]: {
-    uri: string;
-    seriesId: string;
-    episodeTitle: string;
-    player: string;
-    episodeNumber: number;
-  };
 };
 
 export type BrowseStackBrowseScreenProps = NativeStackScreenProps<
   BrowseStackParamList,
   BrowseStackScreenNames.Browse
->;
-export type BrowseStackSeriesStackScreenProps = NativeStackScreenProps<
-  BrowseStackParamList,
-  BrowseStackScreenNames.SeriesStack
->;
-export type BrowseStackNativePlayerScreenProps = NativeStackScreenProps<
-  BrowseStackParamList,
-  BrowseStackScreenNames.NativePlayer
 >;
