+++ apps/mobile/src/routes/main/index.ts
@@ -1,9 +1,11 @@
 export * from './defaultSubHeaderOptions';
 export * from './defaultHeaderOptions';
-export * from './root.interfaces';
-export * from './RootStack';
+export * from './bottom-tab.interfaces';
+export * from './BottomTabStack';
 export * from './browse';
 export * from './mylist';
 export * from './series';
 export * from './settings';
 export * from './search';
+export * from './root.interfaces';
+export * from './RootStack';
