+++ apps/mobile/src/routes/main/mylist/offline/interface.ts
@@ -4,13 +4,11 @@ import { IOfflineSeries } from '../../../../services';
 export enum DownloadStackScreenNames {
   List = 'List',
   SeriesEpisodes = 'SeriesEpisodes',
-  WatchNative = 'WatchNative',
 }
 
 export type DownloadStackParamList = {
   [DownloadStackScreenNames.List]: undefined;
   [DownloadStackScreenNames.SeriesEpisodes]: IOfflineSeries;
-  [DownloadStackScreenNames.WatchNative]: undefined;
 };
 
 export type DownloadStackListScreenProps = NativeStackScreenProps<
@@ -21,9 +19,4 @@ export type DownloadStackListScreenProps = NativeStackScreenProps<
 export type DownloadStackSeriesEpisodesScreenProps = NativeStackScreenProps<
   DownloadStackParamList,
   DownloadStackScreenNames.SeriesEpisodes
->;
-
-export type DownloadStackWatchNativeScreenProps = NativeStackScreenProps<
-  DownloadStackParamList,
-  DownloadStackScreenNames.WatchNative
 >;
